Made AutoAligner work with turret, where it returns the max the turret can turn in the closest direction to the goal, and also the additional degrees the drivetrain needs to turn to compensate for turret limitations.
For BaseAuto, the auto aim no longer turns the drivetrain, so the paths need to turn the robot into a viable heading when shooting. It always autoaligns the hood as much as it can though.

For BaseTeleOp, the hood will also always autoalign, and the drivetrain + shooter + hood will only go to the correct headings/velocities when the driver presses the auto align button. There is a manual override where the turret can be controlled manually by controller 2, but that requires a button to be pressed on controller 1.

Also fixed ratio for shooter, and added a twelve artifact far blue auton that opens the gate, and definitely needs a lot of testing.